Initial setup of metronome fields (beats per measure, unit of measurement, BPM) and Start/Stop buttons.
Adding basically functionality for playing beat sounds via Javscript Audio API.

The Switcher class is used to get around a seeming limitation of playing multiple sounds in a row. Each Channel is loaded with either a Beat or Accent audio file, and must be paused to release audio playback for the next beat to be played. The Switcher loads multiple channels to enable consistent playback of beats when triggered, without having to pause, reinitialize a new Channel, and play the audio again.